langchain 2/requirements.txt
autodoc_pydantic==1.8.0
myst_parser
nbsphinx==0.8.9
sphinx==4.5.0
sphinx-autobuild==2021.3.14
sphinx_book_theme
sphinx_rtd_theme==1.0.0
sphinx-typlog-theme==0.8.0
sphinx-panels
toml
myst_nb
pytest
openai
fastapi
spacy
nltk
uvicorn
pyyaml
itsdangerous
authlib
httpx
starlette[full]
google.auth
transformers
google
google-api-python-client
google-auth-httplib2
google-auth-oauthlib